+++ src/mui/detail/Edit.js
@@ -1,11 +1,10 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { Card, CardTitle, CardActions } from 'material-ui/Card';
+import { Card, CardTitle, CardText } from 'material-ui/Card';
 import inflection from 'inflection';
 import Title from '../layout/Title';
-import { ListButton, DeleteButton, ShowButton } from '../button';
 import { crudGetOne as crudGetOneAction, crudUpdate as crudUpdateAction } from '../../actions/dataActions';
-import RecordForm from './RecordForm'; // eslint-disable-line import/no-named-as-default
+import DefaultActions from './EditActions';
 
 /**
  * Turns a children data structure (either single child or array of children) into an array.
@@ -16,20 +15,27 @@ const arrayizeChildren = children => (Array.isArray(children) ? children : [chil
 export class Edit extends Component {
     constructor(props) {
         super(props);
-        this.state = { record: props.data };
+        this.state = {
+            key: 0,
+            record: props.data,
+        };
         this.handleSubmit = this.handleSubmit.bind(this);
     }
 
     componentDidMount() {
-        this.props.crudGetOne(this.props.resource, this.props.id, this.getBasePath());
+        this.updateData();
     }
 
     componentWillReceiveProps(nextProps) {
         if (this.props.data !== nextProps.data) {
             this.setState({ record: nextProps.data }); // FIXME: erases user entry when fetch response arrives late
+            if (this.fullRefresh) {
+                this.fullRefresh = false;
+                this.setState({ key: this.state.key + 1 });
+            }
         }
         if (this.props.id !== nextProps.id) {
-            this.props.crudGetOne(nextProps.resource, nextProps.id, this.getBasePath());
+            this.updateData(nextProps.resource, nextProps.id);
         }
     }
 
@@ -50,39 +56,51 @@ export class Edit extends Component {
         return location.pathname.split('/').slice(0, -1).join('/');
     }
 
+    updateData(resource = this.props.resource, id = this.props.id) {
+        this.props.crudGetOne(resource, id, this.getBasePath());
+    }
+
+    refresh = (event) => {
+        event.stopPropagation();
+        this.fullRefresh = true;
+        this.updateData();
+    }
+
     handleSubmit(record) {
         this.props.crudUpdate(this.props.resource, this.props.id, record, this.getBasePath());
     }
 
     render() {
-        const { title, children, id, data, isLoading, resource, hasDelete, hasShow, validation } = this.props;
+        const { actions = <DefaultActions />, children, data, hasDelete, hasShow, id, isLoading, resource, title } = this.props;
+        const { key } = this.state;
         const basePath = this.getBasePath();
 
         return (
-            <Card style={{ margin: '2em', opacity: isLoading ? 0.8 : 1 }}>
-                <CardActions style={{ zIndex: 2, display: 'inline-block', float: 'right' }}>
-                    {hasShow && <ShowButton basePath={basePath} record={data} />}
-                    <ListButton basePath={basePath} />
-                    {hasDelete && <DeleteButton basePath={basePath} record={data} />}
-                </CardActions>
-                <CardTitle title={<Title title={title} record={data} defaultTitle={`${inflection.humanize(inflection.singularize(resource))} #${id}`} />} />
-                {data && <RecordForm
-                    onSubmit={this.handleSubmit}
-                    record={data}
-                    resource={resource}
-                    basePath={basePath}
-                    initialValues={data}
-                    validation={validation}
-                >
-                    {children}
-                </RecordForm>}
+            <Card style={{ margin: '2em', opacity: isLoading ? 0.8 : 1 }} key={key}>
+                {actions && React.cloneElement(actions, {
+                    basePath,
+                    data,
+                    hasDelete,
+                    hasShow,
+                    refresh: this.refresh,
+                    resource,
+                })}
+                {data && <CardTitle title={<Title title={title} record={data} defaultTitle={`${inflection.humanize(inflection.singularize(resource))} #${id}`} />} />}
+                {data && React.cloneElement(children, {
+                    onSubmit: this.handleSubmit,
+                    resource,
+                    basePath,
+                    record: data,
+                })}
+                {!data && <CardText>&nbsp;</CardText>}
             </Card>
         );
     }
 }
 
 Edit.propTypes = {
-    children: PropTypes.node,
+    actions: PropTypes.element,
+    children: PropTypes.element.isRequired,
     crudGetOne: PropTypes.func.isRequired,
     crudUpdate: PropTypes.func.isRequired,
     data: PropTypes.object,