+++ src/mui/field/ChipField.js
@@ -2,13 +2,19 @@ import React, { PropTypes } from 'react';
 import Chip from 'material-ui/Chip';
 import get from 'lodash.get';
 
-const ChipField = ({ source, record = {}, style = { margin: 4 } }) => <Chip style={style}>{get(record, source)}</Chip>;
+const ChipField = ({ source, record = {}, elStyle = { margin: 4 } }) =>
+    <Chip style={elStyle}>{get(record, source)}</Chip>;
 
 ChipField.propTypes = {
+    addLabel: PropTypes.bool,
+    elStyle: PropTypes.object,
     label: PropTypes.string,
     source: PropTypes.string.isRequired,
     record: PropTypes.object,
-    style: PropTypes.object,
+};
+
+ChipField.defaultProps = {
+    addLabel: true,
 };
 
 export default ChipField;