+++ src/mui/field/UrlField.js
@@ -1,16 +1,22 @@
 import React, { PropTypes } from 'react';
 import get from 'lodash.get';
 
-const UrlField = ({ source, record = {}, style }) => (
-    <a href={get(record, source)} style={style}>
+const UrlField = ({ source, record = {}, elStyle }) => (
+    <a href={get(record, source)} style={elStyle}>
         {get(record, source)}
     </a>
 );
 
 UrlField.propTypes = {
+    addLabel: PropTypes.bool,
+    elStyle: PropTypes.object,
+    label: PropTypes.string,
     record: PropTypes.object,
     source: PropTypes.string.isRequired,
-    style: PropTypes.object,
+};
+
+UrlField.defaultProps = {
+    addLabel: true,
 };
 
 export default UrlField;