+++ src/mui/input/TextInput.js
@@ -2,34 +2,48 @@ import React, { PropTypes } from 'react';
 import TextField from 'material-ui/TextField';
 import title from '../../util/title';
 
-const TextInput = ({ input, label, meta: { touched, error }, options, type, source, style }) => (
+/**
+ * An Input component for a string
+ *
+ * @example
+ * <TextInput source="first_name" />
+ *
+ * You can customize the `type` props (which defaults to "text").
+ * Note that, due to a React bug, you should use `<NumberField>` instead of using type="number".
+ * @example
+ * <TextInput source="email" type="email" />
+ * <NumberInput source="nb_views" />
+ *
+ * The object passed as `options` props is passed to the material-ui <TextField> component
+ */
+const TextInput = ({ input, label, meta: { touched, error }, options, type, source, elStyle }) => (
     <TextField
         value={input.value}
         onChange={input.onChange}
         type={type}
         floatingLabelText={title(label, source)}
         errorText={touched && error}
-        style={style}
+        style={elStyle}
         {...options}
     />
 );
 
 TextInput.propTypes = {
-    includesLabel: PropTypes.bool.isRequired,
+    addField: PropTypes.bool.isRequired,
+    elStyle: PropTypes.object,
     input: PropTypes.object,
     label: PropTypes.string,
     meta: PropTypes.object,
     name: PropTypes.string,
     onChange: PropTypes.func,
     options: PropTypes.object,
-    source: PropTypes.string.isRequired,
-    style: PropTypes.object,
+    source: PropTypes.string,
     type: PropTypes.string,
     validation: PropTypes.oneOfType([PropTypes.object, PropTypes.func])
 };
 
 TextInput.defaultProps = {
-    includesLabel: true,
+    addField: true,
     options: {},
     type: 'text',
 };