+++ src/mui/list/FilterForm.js
@@ -22,21 +22,24 @@ export class FilterForm extends Component {
         const { currentFilters, resource, style } = this.props;
         return (<div>
             <CardText style={{ float: 'right', marginTop: '-14px', paddingTop: 0, ...style }}>
-                {this.getShownFilters().map(filterElement =>
-                    <div key={filterElement.props.source} style={filterElement.props.style}>
+                {this.getShownFilters().reverse().map(filterElement =>
+                    <div key={filterElement.props.source} style={filterElement.props.style || { display: 'inline-block' }}>
                         {filterElement.props.alwaysOn ?
                             <div style={{ width: 48, display: 'inline-block' }}>&nbsp;</div> :
                             <IconButton iconStyle={{ color: '#00bcd4' }} onTouchTap={this.handleHide} data-key={filterElement.props.source} tooltip="Remove this filter">
                                 <ActionHide />
                             </IconButton>
                         }
-                        <Field
-                            {...filterElement.props}
-                            name={filterElement.props.source}
-                            component={filterElement.type}
-                            resource={resource}
-                            record={currentFilters}
-                        />
+                        <div style={{ display: 'inline-block' }}>
+                            <Field
+                                allowEmpty
+                                {...filterElement.props}
+                                name={filterElement.props.source}
+                                component={filterElement.type}
+                                resource={resource}
+                                record={currentFilters}
+                            />
+                        </div>
                     </div>
                 )}
             </CardText>